Remove renamed file.
