+++ dotnet/src/Functions/Functions.UnitTests/OpenApi/Extensions/OpenApiKernelExtensionsTests.cs
@@ -244,6 +244,23 @@ public sealed class OpenApiKernelExtensionsTests : IDisposable
         Assert.True(plugin.TryGetFunction("IssuesCreatemilestone", out var _));
     }
 
+    [Fact]
+    public async Task ItCanIncludeOpenApiDeleteAndPatchOperationsAsync()
+    {
+        // Arrange
+        var openApiDocument = ResourcePluginsProvider.LoadFromResource("repair-service.json");
+
+        // Act
+        var plugin = await this._kernel.ImportPluginFromOpenApiAsync("repairServicePlugin", openApiDocument, this._executionParameters);
+
+        // Assert
+        Assert.NotNull(plugin);
+        var functionsMetadata = plugin.GetFunctionsMetadata();
+        Assert.Equal(4, functionsMetadata.Count);
+        AssertPayloadParameters(plugin, "updateRepair");
+        AssertPayloadParameters(plugin, "deleteRepair");
+    }
+
     [Theory]
     [InlineData("documentV2_0.json")]
     [InlineData("documentV3_0.json")]
@@ -291,6 +308,15 @@ public sealed class OpenApiKernelExtensionsTests : IDisposable
     }
 
     #region private ================================================================================
+
+    private static void AssertPayloadParameters(KernelPlugin plugin, string functionName)
+    {
+        Assert.True(plugin.TryGetFunction(functionName, out var function));
+        Assert.NotNull(function.Metadata.Parameters);
+        Assert.Equal(2, function.Metadata.Parameters.Count);
+        Assert.Equal("payload", function.Metadata.Parameters[0].Name);
+        Assert.Equal("content_type", function.Metadata.Parameters[1].Name);
+    }
 
     private KernelArguments GetFakeFunctionArguments()
     {
